+++ api/tests/team/unit/application/organization-invitation/organization-invitation.controller.test.js
@@ -1,7 +1,9 @@
 import { MissingQueryParamError } from '../../../../../lib/application/http-errors.js';
+import { usecases as libUsecases } from '../../../../../lib/domain/usecases/index.js';
 import { organizationInvitationController } from '../../../../../src/team/application/organization-invitations/organization-invitation.controller.js';
+import { OrganizationInvitation } from '../../../../../src/team/domain/models/OrganizationInvitation.js';
 import { usecases } from '../../../../../src/team/domain/usecases/index.js';
-import { catchErr, expect, hFake, sinon } from '../../../../../tests/test-helper.js';
+import { catchErr, domainBuilder, expect, hFake, sinon } from '../../../../../tests/test-helper.js';
 
 describe('Unit | Team | Application | Controller | organization-invitation', function () {
   describe('#getOrganizationInvitation', function () {
@@ -44,6 +46,35 @@ describe('Unit | Team | Application | Controller | organization-invitation', fun
 
       // then
       expect(errorCatched).to.be.instanceof(MissingQueryParamError);
+    });
+  });
+
+  describe('#cancelOrganizationInvitation', function () {
+    it('calls the use case to cancel invitation with organizationInvitationId', async function () {
+      //given
+      const organizationInvitationId = 123;
+
+      const request = {
+        auth: { credentials: { userId: 1 } },
+        params: { organizationInvitationId },
+      };
+      const cancelledOrganizationInvitation = domainBuilder.buildOrganizationInvitation({
+        id: organizationInvitationId,
+        status: OrganizationInvitation.StatusType.CANCELLED,
+      });
+
+      sinon
+        .stub(libUsecases, 'cancelOrganizationInvitation')
+        .withArgs({
+          organizationInvitationId: cancelledOrganizationInvitation.id,
+        })
+        .resolves(cancelledOrganizationInvitation);
+
+      // when
+      const response = await organizationInvitationController.cancelOrganizationInvitation(request, hFake);
+
+      // then
+      expect(response.statusCode).to.equal(204);
     });
   });
 });
